chunktoblock.frag/vert: needs higher precision

+++ lib/tfrheightmap/heightmap/update/opengl/chunktoblock.vert
@@ -1,7 +1,7 @@
 attribute highp vec4 qt_Vertex;
 attribute highp vec2 qt_MultiTexCoord0;
 uniform highp mat4 qt_ModelViewProjectionMatrix;
-varying lowp vec2 qt_TexCoord0;
+varying highp vec2 qt_TexCoord0;
 
 void main(void)
 {
